// src/app/modules/lshp-admin/models/institute-licence-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InstituteLicenceView
 */
export interface InstituteLicenceView { 
     /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or Sets licenceStatus
     */
    licenseStatusCode: string;
}
  