// src/app/modules/lshp-admin/models/institution-license-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InstitutionLicenseView
 */
export interface InstitutionLicenseView { 
    /**
     * Gets or sets
     */
    numberOfSchools: number;
        /**
     * Gets or sets
     */
    billingCycleId: number;
    /**
     * Gets or sets
     */
    billingStartDate: any;
    /**
     * Gets or sets
     */
    billingAmount: number;
    /**
     * Gets or sets StatusId
     */
    statusId?: number;

    /**
     * Gets or sets InstituteId
     */
    instituteId?: string;

 
}