// src/app/modules/lshp-admin/models/license-details-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InstituteLicenceView
 */
export interface LicenseListView { 
    /**
     * Gets or sets LicenceStatus
     */
    id?: string;
    billingCycleId?: Array<string>;
    numberOfSchools?: Array<string>;
    billingAmount?: Array<string>;
    billingCycle?: any;
    billingStartDate: any;
    billingStartDateBegin?: string;
    billingStartDateEnd?: string;
    statusId: Array<number>;
    statusCode: Array<string>;
    licenseStatus?: any;
    sortBy: string;
    sortOrder?: any;
    pageNumber?: number;
    pageSize?: number;
   
}